refactor(chat): extract thinking/generating indicator components from assistant.tsx
assistant.tsx was a 1471-line file mixing message rendering with the
turn-waiting UI (ThinkingState, RunStatusIndicator, GeneratingFooter,
ThoughtSummary). That cluster only reads its own props plus
useOptionalChatStream/useClockTick, so it moves to its own file
byte-for-byte with no behavior change.
